+++ CHANGELOG.md
@@ -24,3 +24,4 @@
   - feature: new `requirement` property for `command alias` configuration, you can now specify the `command requirement` explicitly.
     - Before: the `command requirement` for an `alias command` is inferred implicitly from the target command.
     - Now: You can specify the `command requirement` for an `alias command` explicitly, making it clear.
+  - feature: new `document` property for `command alias` configuration, now you can specify the `document string` for each alias command.